+++ sdk/monitor/monitor-opentelemetry-exporter/src/types.ts
@@ -170,6 +170,14 @@ export const legacySemanticValues = [
 ];
 
 /**
+ * Experimental OpenTelemetry semantic convention values
+ * @internal
+ */
+export enum experimentalOpenTelemetryValues {
+  SYNTHETIC_TYPE = "user_agent.synthetic.type",
+}
+
+/**
  * HTTP semantic convention values
  * @internal
  */
@@ -196,7 +204,14 @@ export const httpSemanticValues = [
   ATTR_EXCEPTION_TYPE,
   ATTR_EXCEPTION_MESSAGE,
   ATTR_EXCEPTION_STACKTRACE,
+  experimentalOpenTelemetryValues.SYNTHETIC_TYPE,
 ];
+
+/**
+ * Synthetic source values
+ * @internal
+ */
+export const syntheticSourceValues = ["bot", "test"];
 
 /**
  * Internal Microsoft attributes